+++ lib/src/logic/field_cubit/field_cubit.dart
@@ -1,12 +1,12 @@
 import 'package:bloc/bloc.dart';
 import 'package:cubit_form/src/logic/models/validation.dart';
-import 'package:flutter/foundation.dart';
+import 'package:equatable/equatable.dart';
 
 part 'field_cubit_state.dart';
 
 class FieldCubit<T> extends Cubit<FieldCubitState<T>> {
   FieldCubit({
-    @required this.initalValue,
+    required this.initalValue,
     this.validations = const [],
   }) : super(FieldCubitState<T>(
           value: initalValue,
@@ -17,19 +17,23 @@ class FieldCubit<T> extends Cubit<FieldCubitState<T>> {
   final T initalValue;
   final List<ValidationModel<T>> validations;
 
-  void setValue(T value) => emit(
-        state.copyWith(
-          value: value,
-          error: tillfirstError<T>(value, validations),
-        ),
-      );
+  void setValue(T value) {
+    return emit(
+      state.copyWith(
+        value: value,
+        error: tillfirstError<T>(value, validations),
+      ),
+    );
+  }
 
-  void externalSetValue(T value) => emit(
-        state.externalChange(
-          value: value,
-          error: tillfirstError<T>(value, validations),
-        ),
-      );
+  void externalSetValue(T value) {
+    return emit(
+      state.externalChange(
+        value: value,
+        error: tillfirstError<T>(value, validations),
+      ),
+    );
+  }
 
   //for async validation
   void setError(String error) => emit(
@@ -59,8 +63,8 @@ class FieldCubit<T> extends Cubit<FieldCubitState<T>> {
       );
 }
 
-String tillfirstError<T>(T value, validations) {
-  String error;
+String? tillfirstError<T>(T value, validations) {
+  String? error;
   for (var validation in validations) {
     error = validation.check(value);
     if (error != null) {